meeting changes
Arm control isn't quite working yet. I think we might want it to switch
to stabilizing the winch when the forearm is close to the target. It
also might be better to switch to polar control for IK, since it will
mostly work even when the shoulder gear slips. The sensor to angle
calculation also needs some tuning. I will try to get control that
works independently of the absolute encoder position tonight in the
simulator.